added septrate texture file
