+++ map/map1.cub
@@ -8,7 +8,7 @@ NO 				            ./texture/stone_wall.png
 	
 		EA   	  ./texture/brick.png 		    
  
- WE     	  ./texture/standstone.png
+ WE     	  ./texture/x.png
 
 
  	 	 	 	 F      27,32,11
